Refacto MNSIT tests
